[FIX] FeedTotalsResponse totals 타입을 Long으로 수정

+++ src/main/java/com/study/api_gateway/dto/activity/response/FeedTotalsResponse.java
@@ -12,6 +12,6 @@ import java.util.Map;
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class FeedTotalsResponse {
-	private Map<String, Integer> totals;
+	private Map<String, Long> totals;
 	private Boolean isOwner;
 }
